+++ public/style.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700;900&display=swap');
-
 :root {
 	--primary-color: #fd951f;
 	--secondary-color: #1f83fd;
@@ -178,6 +176,7 @@ form .item > div:nth-child(1) {
 	margin-bottom: 8px;
 }
 
+form .item .input input,
 form .item input[type='text'],
 form .item input[type='number'],
 form .item textarea,
@@ -509,4 +508,17 @@ form .item #photos-preview img {
 .search-products nav ul li a:hover {
 	color: black;
 	font-weight: bold;
+}
+
+/*! USER REGISTER */
+.user-register.container.form {
+	max-width: 500px;
+}
+
+.form input[name='email'],
+.form input[name='password'],
+.form input[name='passwordRepeat'],
+.form input[name='cpf_cnpj'],
+.form input[name='cep'] {
+	max-width: 60%;
 }
