+++ src/app/controllers/user_controller.js
@@ -1,0 +1,5 @@
+module.exports = {
+	registerForm(req, res) {
+		return res.render('user/register')
+	}
+}
