+++ src/app/models/file.js
@@ -1,4 +1,4 @@
-const db = require('../config/db.js')
+const db = require('../../config/db.js')
 const fs = require('fs')
 
 module.exports = {
