+++ src/app/views/user/fields.html
@@ -1,0 +1,70 @@
+<div class="item">
+	<div>Nome Completo</div>
+	<div class="input">
+		<input
+			type="text"
+			name="name"
+			placeholder="Digite seu nome completo"
+			value="{{ user.name }}"
+		/>
+	</div>
+</div>
+
+<div class="item">
+	<div>Email</div>
+	<div class="input">
+		<input type="email" name="email" placeholder="Digite seu email" value="{{ user.email }}" />
+	</div>
+</div>
+
+<div class="item">
+	<div>Senha</div>
+	<div class="input">
+		<input
+			type="password"
+			name="password"
+			placeholder="Escolha uma senha segura"
+			value="{{ user.password }}"
+		/>
+	</div>
+</div>
+
+{% if not user.id %}
+<div class="item">
+	<div>Repetir Senha</div>
+	<div class="input">
+		<input type="password" name="passwordRepeat" placeholder="Repita a senha escolhida" />
+	</div>
+</div>
+{% endif %}
+
+<div class="item">
+	<div>CPF ou CNPJ</div>
+	<div class="input">
+		<input
+			type="text"
+			name="cpf_cnpj"
+			placeholder="Digite seu documento"
+			value="{{ user.cpf_cnpj }}"
+		/>
+	</div>
+</div>
+
+<div class="item">
+	<div>CEP</div>
+	<div class="input">
+		<input type="text" name="cep" placeholder="99999-999" value="{{ user.cep }}" />
+	</div>
+</div>
+
+<div class="item">
+	<div>Endereço Completo</div>
+	<div class="input">
+		<input
+			type="text"
+			name="address"
+			placeholder="Digite seu endereço completo"
+			value="{{ user.address }}"
+		/>
+	</div>
+</div>
