+++ src/app/views/user/register.html
@@ -1,0 +1,15 @@
+{% extends 'layout.html' %} {% block content %}
+
+<section class="user-register container form">
+	<form action="/users/register" method="POST" >
+		<div class="header">
+			<h3>Registre-se</h3>
+		</div>
+
+		<div class="fields">{% include 'user/fields.html' %}</div>
+
+		<button class="button" type="submit">Salvar</button>
+	</form>
+</section>
+
+{% endblock %}
