+++ src/server.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const nunjucks = require('nunjucks')
-const routes = require('./routes.js')
+const routes = require('./routes')
 const methodOverride = require('method-override')
 
 const server = express()
